Make alerts hook handle component creation
