+++ client/src/containers/AddChild/AddChild.tsx
@@ -159,7 +159,15 @@ const AddChild: React.FC = () => {
     if (indexOfCurrentStep === steps.length - 1) {
       // If we're all done
       console.log('TODO: what do we do after adding a child?');
-      history.push('/check-data/1', { alerts: [{ type: 'success', heading: 'Record added', text: `${child?.firstName} ${child?.lastName}'s record was added to your roster.` }] });
+      history.push('/check-data/1', {
+        alerts: [
+          {
+            type: 'success',
+            heading: 'Record added',
+            text: `${child?.firstName} ${child?.lastName}'s record was added to your roster.`,
+          },
+        ],
+      });
     } else {
       setRefetchChild((r) => r + 1);
       history.replace({ ...location, hash: steps[indexOfCurrentStep + 1].key });
