+++ client/src/containers/CheckData/CheckData.tsx
@@ -6,10 +6,10 @@ import 'react-perfect-scrollbar/dist/css/styles.css';
 import AuthenticationContext from '../../contexts/AuthenticationContext/AuthenticationContext';
 import { apiGet } from '../../utils/api';
 import { Child } from '../../shared/models';
-import { Button, Table, Alert } from '@ctoec/component-library';
+import { Button, Table } from '@ctoec/component-library';
 import { tableColumns } from './TableColumns';
 import { BackButton } from '../../components/BackButton';
-import useAlerts from '../../hooks/useAlertsAfterNavigation';
+import useAlerts from '../../hooks/useAlerts';
 
 const CheckData: React.FC = () => {
   const { reportId } = useParams();
@@ -40,7 +40,7 @@ const CheckData: React.FC = () => {
       <div className="CheckData__content margin-top-4 grid-container">
         <div className="margin-x-4">
           <BackButton />
-          {alerts.map((alertProps, i) => <Alert {...alertProps} key={i} />)}
+          {alerts}
           <h1>Check data for {pluralize('child', reportData.length, true)}</h1>
           <p>Make sure all of your data was uploaded correctly. </p>
           <p>If everything looks good, submit to OEC.</p>
@@ -57,8 +57,8 @@ const CheckData: React.FC = () => {
               />
             </PerfectScrollbar>
           ) : (
-              'Loading...'
-            )}
+            'Loading...'
+          )}
         </div>
       </div>
       <div className="CheckData__button-container position-fixed bottom-0 width-full">
